Hopefully fixed multiplayer: update 0.10.7a
