+++ .classpath
@@ -62,5 +62,6 @@
 	<classpathentry exported="true" kind="lib" path="C:/Users/Nathan/.gradle/caches/modules-2/files-2.1/com.google.guava/guava/17.0/9c6ef172e8de35fd8d4d8783e4821e57cdef7445/guava-17.0.jar" sourcepath="C:/Users/Nathan/.gradle/caches/modules-2/files-2.1/com.google.guava/guava/17.0/7ca0efbeb87ca845b5d7a0ac9c21a4b7b95f7b28/guava-17.0-sources.jar"/>
 	<classpathentry exported="true" kind="lib" path="C:/Users/Nathan/.gradle/caches/modules-2/files-2.1/org.apache.commons/commons-lang3/3.3.2/90a3822c38ec8c996e84c16a3477ef632cbc87a3/commons-lang3-3.3.2.jar" sourcepath="C:/Users/Nathan/.gradle/caches/modules-2/files-2.1/org.apache.commons/commons-lang3/3.3.2/d2a489573c0ed2c4942b3660decad5d65087b406/commons-lang3-3.3.2-sources.jar"/>
 	<classpathentry exported="true" kind="lib" path="C:/Users/Nathan/.gradle/caches/minecraft/net/minecraftforge/forge/1.7.10-10.13.2.1230/start"/>
+	<classpathentry kind="lib" path="libs/Railcraft_1.7.10-9.5.0.0-dev.jar"/>
 	<classpathentry kind="output" path="bin"/>
 </classpath>
